generative-ai/llm/3-practice/3-3-customer-support-rag/src/rag/chains/chain.py
from langchain_core.runnables import RunnablePassthrough
from langchain_core.output_parsers import StrOutputParser
from prompts.prompt_template import prompt


def init_rag_chain(llm, retriever):
    return (
        {
            "context": retriever,
            "input": RunnablePassthrough()
        }
        | prompt
        | llm
        | StrOutputParser()
    )