generative-ai/llm/3-practice/3-3-customer-support-rag/pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.2.9",
    "langchain>=0.3.8",
    "langchain-community>=0.3.8",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.1",
    "jq>=1.8.0",
    "langchain-chroma>=0.1.4",
]
